URL Cleanup (#7)
This commit updates URLs to prefer the https protocol. Redirects are not followed to avoid accidentally expanding intentionally shortened URLs (i.e. if using a URL shortener).

# HTTP URLs that Could Not Be Fixed
These URLs were unable to be fixed. Please review them to see if they can be manually resolved.

- [x] http://my-website.com/path/to/image.jpg (200) with 1 occurrences could not be migrated:  
   ([https](https://my-website.com/path/to/image.jpg) result ConnectTimeoutException). **manually migrated: example in comment describing payload format**

# Fixed URLs

## Fixed But Review Recommended
These URLs were fixed, but the https status was not OK. However, the https status was the same as the http request or http redirected to an https URL, so they were migrated. Your review is recommended.

* http://flickr.com/icons/bobby.jpg (403) with 1 occurrences migrated to:  
  https://flickr.com/icons/bobby.jpg ([https](https://flickr.com/icons/bobby.jpg) result 403). **ok: example in comment describing payload format**
* http://example.com/path/to/thumb.png (404) with 1 occurrences migrated to:  
  https://example.com/path/to/thumb.png ([https](https://example.com/path/to/thumb.png) result 404). **ok: example in comment describing payload format**

## Fixed Success 
These URLs were switched to an https URL with a 2xx status. While the status was successful, your review is still recommended.

* http://projectreactor.io with 4 occurrences migrated to:  
  https://projectreactor.io ([https](https://projectreactor.io) result 200).
* http://flickr.com/bobby/ with 1 occurrences migrated to:  
  https://flickr.com/bobby/ ([https](https://flickr.com/bobby/) result 302).