Distinguish specific Repo config and common config
This commit allows the triage label and forward-merge hint use cases to
be blanket-enabled via a common configuration (repo name == `*`).

The resolution of the common config is made more explicit, and the
resulting Repo (which materializes the configuration) is synthetically
made from the common Repo (so that it has a valid name).

The forward-merge hint was previously not issued to repositories without
an exact match in Repo configurations, because the comment is made to
Repo.repo, which was inexisting `*` prior to this change.