Activate forward-merge hint on *.x branches via opt-in (#15)
This commit enables the forward-merge hint comment on PR close without
an exact Repo subconfiguration. If a Repo exactly matches the PR's, then
the list of maintainers may be used as a fallback for when no merge
information has been provided.

Nonetheless, the main way of activating the feature is to opt-in into it
via the new github.mergeHintRepos list property. Any PR targeting a repo
that isn't in this list will be ignored with 204. It should be safe to
assume that a PR merge would always have merger information, and only
authorize team members can merge anyway, so the Repo configuration isn't
super useful.

Additional polish:

The resolution of repo config now explicitly only look for an exact
match. The methods that look for a common Repo config (with name "*" are
now more explicit about that fact.

The forward-merge hint was previously not issued to repositories without
an exact match in Repo configurations, because the comment is made to
Repo.repo, which was inexisting `*` prior to this change. The order in
which PR webhook criterias are inspected has been reworked.

Tests around the configuration have been added, and the order in which
the property sources are evaluated has been revised so that the custom
property file is injected between application.properties and a potential
application-test.properties:
 - application-test.properties takes precedence in unit tests that have
   the profile active
 - .reactor-bot/reactor-bot.properties allows to run the bot locally and
   perform manual integration testing
 - application.properties has the default base values that aren't secret